requirements.txt
tensorboardX
opencv-contrib-python
pathlib
scikit-image
pandas
numpy
mmcv
torchvision
albumentations==0.3.1
terminaltables
tqdm
